// server/vance-addon-brain-workbook/client/src/generated/workbook/WorkbookView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.workbook.WorkbookView
*/

import { WorkbookPageView } from './WorkbookPageView';

export interface WorkbookView {
  folder: string;
  title?: string;
  description?: string;
  landingPagePath?: string;
  landingPageId?: string;
  indexPagePath?: string;
  indexPageId?: string;
  pages: WorkbookPageView[];
}